+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/NullableCheckBox.java
@@ -135,7 +135,7 @@ public class NullableCheckBox extends JCheckBox {
 		@Override
 		public void mouseClicked(MouseEvent e) {
 			if (isEnabled() && (e == null || notModified(e))) {
-				model().toggleState().next();
+				model().state().next();
 			}
 		}
 
@@ -152,7 +152,7 @@ public class NullableCheckBox extends JCheckBox {
 		@Override
 		public void paintIcon(Component component, Graphics graphics, int x, int y) {
 			icon.paintIcon(component, graphics, x, y);
-			if (model().toggleState().get() != null) {
+			if (model().state().get() != null) {
 				return;
 			}
 
@@ -195,7 +195,7 @@ public class NullableCheckBox extends JCheckBox {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
-			model.toggleState().next();
+			model.state().next();
 		}
 	}
 }